Moves metrics to metrics file
